Move maps to a dedicated module
